# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iFal API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register Users",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Ivanova"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "a.ivanova123@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "pass123"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "+380661234567"
                                    },
                                    "ageGroup": {
                                        "type": "string",
                                        "example": "adult"
                                    },
                                    "birthDate": {
                                        "type": "string",
                                        "example": "1990-06-15"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "female"
                                    },
                                    "loginType": {
                                        "type": "string",
                                        "example": "NORMAL"
                                    },
                                    "countryCode": {
                                        "type": "string",
                                        "example": "UA"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "student"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created User object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "firstname": {
                                            "type": "string",
                                            "example": "Anna"
                                        },
                                        "lastname": {
                                            "type": "string",
                                            "example": "Ivanova"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "a.ivanova123@gmail.com"
                                        },
                                        "profile_image": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "is_active": {
                                            "type": "boolean",
                                            "nullable": true
                                        },
                                        "deleted_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "email_verified_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "remember_token": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "role": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "student"
                                                    },
                                                    "guard_name": {
                                                        "type": "string",
                                                        "example": "web"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-09-11T07:16:56.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-09-11T07:16:56.000000Z"
                                                    },
                                                    "pivot": {
                                                        "properties": {
                                                            "model_id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "role_id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "model_type": {
                                                                "type": "string",
                                                                "example": "App\\\\Models\\\\User"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login Users",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "a.ivanova123@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "pass123"
                                    },
                                    "loginType": {
                                        "type": "string",
                                        "example": "NORMAL"
                                    },
                                    "deviceType": {
                                        "type": "string",
                                        "example": "tablet",
                                        "nullable": true
                                    },
                                    "OSVersion": {
                                        "type": "string",
                                        "example": "16.2",
                                        "nullable": true
                                    },
                                    "deviceToken": {
                                        "type": "string",
                                        "example": "mbiDmobMRlhnDYMZa9",
                                        "nullable": true
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Apple",
                                        "nullable": true
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "iPhone 11",
                                        "nullable": true
                                    },
                                    "deviceId": {
                                        "type": "string",
                                        "example": "5893590235803",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in User object and Created API Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|laravel_sanctum_hBEi8VPgE28bXfmYhjjRcF0stKod4Aqn6lkt6egW4945d232"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/request-verification": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request a verification code",
                "description": "Sends a verification code to the given phone number",
                "operationId": "sendVerificationCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "+380685469492"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Verification code has been sent"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The phone number field format is invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/verify-code": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify the code sent to the phone",
                "description": "Verifies the code sent to the phone number",
                "operationId": "verifyCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "code"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "+380689269492"
                                    },
                                    "code": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phone number verified successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid code"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out",
                "description": "Logs out the current user by invalidating the token",
                "operationId": "logoutUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "firstname": {
                        "type": "string",
                        "example": "Anna"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Ivanova"
                    },
                    "email": {
                        "type": "string",
                        "example": "a.ivanova123@gmail.com"
                    },
                    "profile_image": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "integer",
                        "example": 1
                    },
                    "deleted_at": {
                        "type": "string",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "nullable": true
                    },
                    "remember_token": {
                        "type": "string",
                        "nullable": true
                    },
                    "role": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}